Review fixes for #1221 -Wsign-compare

+++ src/box/box.cc
@@ -78,9 +78,9 @@ struct recovery *recovery;
  */
 static struct recover_row_ctx {
 	/** How many rows have been recovered so far. */
-	int rows;
+	size_t rows;
 	/** Yield once per 'yield' rows. */
-	int yield;
+	size_t yield;
 } recover_row_ctx;
 
 bool snapshot_in_progress = false;
@@ -88,7 +88,7 @@ static bool box_init_done = false;
 bool is_ro = true;
 
 void
-recover_row_ctx_init(struct recover_row_ctx *ctx, int rows_per_wal)
+recover_row_ctx_init(struct recover_row_ctx *ctx, size_t rows_per_wal)
 {
 	ctx->rows = 0;
 	/**
@@ -257,7 +257,8 @@ static void
 box_check_readahead(int readahead)
 {
 	enum { READAHEAD_MIN = 128, READAHEAD_MAX = 2147483648 };
-	if ((unsigned)readahead < READAHEAD_MIN || (unsigned)readahead > READAHEAD_MAX) {
+	if (readahead < (int) READAHEAD_MIN ||
+	    readahead > (int) READAHEAD_MAX) {
 		tnt_raise(ClientError, ER_CFG, "readahead",
 			  "specified value is out of bounds");
 	}
@@ -1065,7 +1066,7 @@ box_init(void)
 	iproto_init();
 	box_set_listen();
 
-	int rows_per_wal = box_check_rows_per_wal(cfg_geti("rows_per_wal"));
+	size_t rows_per_wal = box_check_rows_per_wal(cfg_geti("rows_per_wal"));
 	enum wal_mode wal_mode = box_check_wal_mode(cfg_gets("wal_mode"));
 	recovery_finalize(recovery, wal_mode, rows_per_wal);
